refactor(server): split server.py into main.py + server.py
server.py was both the process entry point (CLI parsing, uvicorn
boot, parent-process watchdog, instrumented heavy-import waterfall,
FastAPI lifespan + heavy init) and the FastAPI app definition.
Split along the natural seam:

- main.py (new): entry point. Owns the heavy-import try-block,
  apply_resolved_token, env setup, ParentWatchdog, _heavy_init,
  startup-broadcast helpers, lifespan, the FastAPI `app` instance,
  CORS middleware, and the uvicorn boot. Mounts server.py's router.
- server.py: just the endpoints. Exposes a `router: APIRouter`
  with /health, /api/model-info, and /ws. Import-safe and
  process-agnostic — no module-level side effects.

Frontend (electron/ipc/engine.ts existence checks; server.ts spawn
command) updated to invoke main.py. CONTRIBUTING.md, README.md,
pyproject.toml per-file ignores, and a few stale comments tracking
"the heavy try-block in server.py" updated to point at main.py.
